Allow multiple feedback classes on a single atom
For use with the current implementation of transitive chains.  As it is, only the feedback icon on a transitive chain of equations is just for the last equation in the chain.  This allows all of the equations to set their relevant feedback class and then css picks the most incorrect one to display.